+++ CHANGELOG.md
@@ -15,8 +15,6 @@ The following emojis are used to highlight certain changes:
 
 ### Added
 
-- Add support for custom DNSLink resolvers (e.g. to support TLDs like `.eth`, `.crypto`). It is possible to set custom DoH resolvers by setting `RAINBOW_DNSLINK_RESOLVERS` with the same convention as Kubo's [`DNS.Resolvers`](https://github.com/ipfs/kubo/blob/master/docs/config.md#dnsresolvers) ) [#224](https://github.com/ipfs/rainbow/pull/224)
-
 ### Changed
 
 ### Fixed
@@ -27,6 +25,11 @@ The following emojis are used to highlight certain changes:
 
 ### Security
 
+## [1.10.1]
+
+### Added
+
+- Add support for custom DNSLink resolvers (e.g. to support TLDs like `.eth`, `.crypto`). It is possible to set custom DoH resolvers by setting `RAINBOW_DNSLINK_RESOLVERS` with the same convention as Kubo's [`DNS.Resolvers`](https://github.com/ipfs/kubo/blob/master/docs/config.md#dnsresolvers) ) [#224](https://github.com/ipfs/rainbow/pull/224)
 
 ## [1.10.0]